Tema4/requirements.txt
Flask==1.1.1
requests==2.23.0
azure-ai-textanalytics==1.0.0b4
azure-cognitiveservices-speech==1.11.0
